Declare sched_pcore in allocator specific file
As part of this, we need to move the definitions of any helper functions
in corerequest.h which access fields of the sched_pcore or otherwise
depend on its size. This will become more important as we start to add
other allocators that need extra fields in the sche_pcore struct, and we
want the accessor functions (e.g. spc2pcoreid()) to remain as static
inlines so they are fast.

We also need to move our declaration of core_reqeust_data struct into
the allocator specific header as well.

I don't really like the #ifdef around including the allocator specific
header files, but it's the best compromise if we want our accessor
functions to remain in the header file. Altrnatively, we could try and
come up with a way to not need these accessor functions at all, but that
is for another day.
